% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{bic}
\alias{bic}
\title{Calculate approximate BIC for the model}
\usage{
bic(data, groups, pi, beta, dynamic = F)
}
\arguments{
\item{data}{N x J response matrix}

\item{groups}{N length vector of group membership indicators in 1,..,G}

\item{pi}{G x K matrix of group-specific distributions over classes, usually posterior mean}

\item{beta}{J-length list of K x L_j class-specific response distributions, usually posterior mean}

\item{dynamic}{Adjusts number of parameters if this is a DHLC model}
}
\value{
BIC value for model with provided data and parameters
}
\description{
Calculate approximate BIC for the model
}
