% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{checkIdCondition}
\alias{checkIdCondition}
\title{Check necessary condition for identification}
\usage{
checkIdCondition(G, Lj)
}
\arguments{
\item{G}{number of groups}

\item{Lj}{J length vector with number of responses for each question}
}
\value{
The maximum number of classes K that meet the identification condition given G, Lj
}
\description{
Check necessary condition for identification
}
