% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{checkSparsity}
\alias{checkSparsity}
\title{Check for which responses and questions the grouped response matrix is sparse}
\usage{
checkSparsity(aux.data, limit = 0.05)
}
\arguments{
\item{aux.data}{}

\item{limit}{}
}
\value{
prints which responses are spares
}
\description{
Check for which responses and questions the grouped response matrix is sparse
}
