% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAnalysis.R
\name{clean_data}
\alias{clean_data}
\title{Cleans up raw survey data matrix of numerical scores  by 1) converting all missing or not-applicable codes 
to a single response code L_j and 2) converting raw scores (that may not be contiguous) to 1:L_j, where the order of 
the original scores is maintained in the cleaned data}
\usage{
clean_data(raw, na.values = NULL)
}
\arguments{
\item{raw}{N x J matrix of raw numeric scores, where each score in a column corresponds to a different question response}

\item{na.values}{vector of values in raw that should be considered missing or N/A responses.
Any values in the matrix that are negative or NA are automatically included.}
}
\value{
An N x J matrix where each column has values only in 1 to L_j
}
\description{
Cleans up raw survey data matrix of numerical scores  by 1) converting all missing or not-applicable codes 
to a single response code L_j and 2) converting raw scores (that may not be contiguous) to 1:L_j, where the order of 
the original scores is maintained in the cleaned data
}
