% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhlvm.R
\docType{package}
\name{dhlvm}
\alias{dhlvm}
\alias{dhlvm-package}
\title{dhlvm}
\description{
Estimates markov models for categorical time series data described in Munro (2018).
}
\author{
Evan Munro <evan.munro@gmail.com>
}
