% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlcEstimators.R
\name{discreteLDSModel}
\alias{discreteLDSModel}
\title{Old HLC model with J=1}
\usage{
discreteLDSModel(data, eta, v0, s0, tune, K, steps, burn, skip)
}
\arguments{
\item{data}{XxV matrix of counts}

\item{eta}{prior for dirichlet dist for beta}

\item{v0}{prior for invgamma dist for sigma}

\item{s0}{prior rate for invgamma}

\item{tune}{tuning parameter for SGLD step}

\item{K}{number of states}

\item{steps}{number of steps for sampler}

\item{burn}{burn-in iterations for sampler}

\item{skip}{thinning iterations for sampler}
}
\value{
List with posterior samples from 1) theta 2) sigma 3) beta
}
\description{
Old HLC model with J=1
}
