% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteTSEstimators.R
\name{discreteMSFast}
\alias{discreteMSFast}
\title{Runs Gibbs Sampler for markov switching model for discrete time series data with c++ loop}
\usage{
discreteMSFast(data, eta, alpha, K, steps, burn, skip)
}
\arguments{
\item{data}{TimexV data frame of counts of data at time t with value v}

\item{eta}{V-length prior for state-specific multinomial probabilities beta}

\item{alpha}{K-length prior for transition matrix}

\item{K}{Number of states}

\item{steps}{Total number of gibbs sampling iterations}

\item{burn}{Number of iterations to discard}

\item{skip}{Thinning parameter for gibbs sampler}
}
\value{
List with posterior mean of 1) state-specific multinomial probabilities beta 2) transition matrix 3) states
}
\description{
Runs Gibbs Sampler for markov switching model for discrete time series data with c++ loop
}
