% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteTSEstimators.R
\name{ldaModel}
\alias{ldaModel}
\title{Title}
\usage{
ldaModel(data, groups, eta, alpha, steps, burn, skip)
}
\arguments{
\item{data}{}

\item{groups}{}

\item{eta}{}

\item{alpha}{}

\item{steps}{}

\item{burn}{}

\item{skip}{}
}
\description{
Title
}
