% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{plotBetas}
\alias{plotBetas}
\title{Plot class-specific distributions over responses in bar chart}
\usage{
plotBetas(betas, response_codes = NULL, questions = NULL)
}
\arguments{
\item{betas}{J-length list of K x L_j matrices of class-specific distributions over responses}

\item{response_codes}{Optional J length list, each containing a vector of response codes for each question to plot}

\item{questions}{Optional J length vector of question labels to title each graph}
}
\value{
Saves plots in current working directlry
}
\description{
Plot class-specific distributions over responses in bar chart
}
