% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteTSEstimators.R
\name{posteriorMean}
\alias{posteriorMean}
\title{Get posterior mean for set of posterior estimates from sampler}
\usage{
posteriorMean(estimate, out)
}
\arguments{
\item{estimate}{3d array from sampler}

\item{out}{vector of iterations to take mean over for gibbs sampler}
}
\value{
matrix of means of posterior over sampler slices
}
\description{
Get posterior mean for set of posterior estimates from sampler
}
