% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{posteriorMeans}
\alias{posteriorMeans}
\title{Get posterior mean of parameters}
\usage{
posteriorMeans(post)
}
\arguments{
\item{post}{Posterior sample from running MCMC for LDA-S or LDA-DS}
}
\value{
List with attribute containing posterior mean for each parameter in the model
}
\description{
Get posterior mean of parameters
}
