% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampleBeta}
\alias{sampleBeta}
\title{Updates state-specific parameters for multinomial distribution given data, prior, and 
state estimates for each time period}
\usage{
sampleBeta(data, S, eta, K, V, Time)
}
\arguments{
\item{data}{T x V matrix of counts}

\item{S}{T length vector of states}

\item{eta}{Dirichlet prior on beta, parameters for multinomial distribution}

\item{K}{Number of states}

\item{V}{number of discrete options that data can take}

\item{Time}{}
}
\value{
updated estimate of parameter for state-specific multinomial distributions
}
\description{
Updates state-specific parameters for multinomial distribution given data, prior, and 
state estimates for each time period
}
