% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampleBetaSS}
\alias{sampleBetaSS}
\title{Updates state-specific parameters for multinomial distribution given data, prior, and 
state estimates for each response in the data}
\usage{
sampleBetaSS(data, z, eta, K, V, Time)
}
\arguments{
\item{data}{T-length list of Nt vectors of responses (integers in 1:V)}

\item{z}{T-length list of Nt vectors of state assignments for each response}

\item{eta}{V-length vector of prior on beta}

\item{K}{Number of states}

\item{V}{Number of unique outcomes}

\item{Time}{Number of time periods}
}
\value{
a KxV matrix, updated estimate of parameter for state-specific multinomial dist.
}
\description{
Updates state-specific parameters for multinomial distribution given data, prior, and 
state estimates for each response in the data
}
