% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampleState}
\alias{sampleState}
\title{Updates state vector given transition matrix, data, beta, and current estimate of state 
Follows single-move sampler from Albert and Chib (1996)}
\usage{
sampleState(data, P, S, beta, Time, K)
}
\arguments{
\item{data}{Time x V counts (word counts, or survey respondent counts, for example)}

\item{P}{K x K Transition matrix}

\item{beta}{K V-length probability vectors for K multinomial distributions}

\item{Time}{number of time periods}

\item{K}{number of markov states}

\item{S_prev}{Time-length vector of state estimates in 1:K}
}
\value{
Time-length vector of new state estimates
}
\description{
Updates state vector given transition matrix, data, beta, and current estimate of state 
Follows single-move sampler from Albert and Chib (1996)
}
