% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampleTransition}
\alias{sampleTransition}
\title{Updates transition matrix given current estimate of states}
\usage{
sampleTransition(S, alpha, K, Time)
}
\arguments{
\item{S}{Time length vector of states}

\item{alpha}{dirichlet prior for transition matrix columns}

\item{K}{Number of states}

\item{Time}{number of time periods in the data}
}
\value{
KxK state transition probability matrix
}
\description{
Updates transition matrix given current estimate of states
}
