% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampleZ}
\alias{sampleZ}
\title{Updates state assignments for each response based on current estimate for 
state-specific multinomial distributions and state mixtures for each date}
\usage{
sampleZ(data, theta, beta, Time)
}
\arguments{
\item{data}{T-length list of N_t length vectors of response indices for each t}

\item{theta}{TxK matrix of state mixture probabilities for each t}

\item{beta}{KxV matrix of state-specific multinomial probabilities}

\item{Time}{Number of time periods in data}
}
\value{
T-length list of Nt vectors containing state assignments for each response
}
\description{
Updates state assignments for each response based on current estimate for 
state-specific multinomial distributions and state mixtures for each date
}
