% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{topBetas}
\alias{topBetas}
\title{Indicate which elements of the profile specific multinomial distributions have the highest weight}
\usage{
topBetas(beta, identifier, top = 5)
}
\arguments{
\item{beta}{KxV data frame of posterior estimates of}

\item{identifier}{V-length vector of names for each v-index of beta}

\item{top}{number of elements of beta to return}
}
\value{
table with top k responses for each profile-specific multinomial distribution
}
\description{
Indicate which elements of the profile specific multinomial distributions have the highest weight
}
