% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlcEstimators.R
\name{dhlcModel}
\alias{dhlcModel}
\title{Title}
\usage{
dhlcModel(X, groups, eta, v0, s0, tune, steps, burn, skip)
}
\arguments{
\item{X}{}

\item{groups}{}

\item{eta}{}

\item{v0}{}

\item{s0}{}

\item{tune}{}

\item{steps}{}

\item{burn}{}

\item{skip}{}
}
\description{
Title
}
