% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhlvm.R
\docType{package}
\name{dhlvm-package}
\alias{dhlvm}
\alias{dhlvm-package}
\title{dhlvm: Dynamic Hierarchical Latent Variable Models}
\description{
Estimates markov models for categorical time series data described in Munro (2018).
}
\references{
\href{http://evanmunro.ca/files/discreteTS.pdf}{Munro (2018)}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/evanmunro/dhlvm/issues}
}

}
\author{
\strong{Maintainer}: Evan Munro \email{munro@stanford.edu}

}
\keyword{internal}
