% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlcEstimators.R
\name{hlcModel}
\alias{hlcModel}
\title{Title}
\usage{
hlcModel(X, groups, eta, alpha, steps, burn, skip)
}
\arguments{
\item{groups}{}

\item{eta}{}

\item{alpha}{}

\item{steps}{}

\item{burn}{}

\item{skip}{}

\item{data}{}
}
\description{
Title
}
