% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAnalysis.R
\name{plotBetas}
\alias{plotBetas}
\title{Title}
\usage{
plotBetas(betas, response_codes = NULL, questions = NULL)
}
\arguments{
\item{betas}{}

\item{questions}{}

\item{rsponse_codes}{}
}
\description{
Title
}
